+++ xlabuilder/special_ops_test.go
@@ -79,16 +79,16 @@ func TestIota(t *testing.T) {
 	client := getPJRTClient(t)
 	builder := New(t.Name())
 
-	iotaOp := capture(Iota(builder, MakeShape(dtypes.F64, 3, 2), 0)).Test(t)
+	iotaOp := capture(Iota(builder, MakeShape(dtypes.Float32, 3, 2), 0)).Test(t)
 	exec := compile(t, client, capture(builder.Build(iotaOp)).Test(t))
-	got, dims := execArrayOutput[float64](t, client, exec)
-	require.Equal(t, []float64{0, 0, 1, 1, 2, 2}, got)
+	got, dims := execArrayOutput[float32](t, client, exec)
+	require.Equal(t, []float32{0, 0, 1, 1, 2, 2}, got)
 	require.Equal(t, []int{3, 2}, dims)
 
-	iotaOp = capture(Iota(builder, MakeShape(dtypes.F64, 2, 3), 1)).Test(t)
+	iotaOp = capture(Iota(builder, MakeShape(dtypes.Float32, 2, 3), 1)).Test(t)
 	exec = compile(t, client, capture(builder.Build(iotaOp)).Test(t))
-	got, dims = execArrayOutput[float64](t, client, exec)
-	require.Equal(t, []float64{0, 1, 2, 0, 1, 2}, got)
+	got, dims = execArrayOutput[float32](t, client, exec)
+	require.Equal(t, []float32{0, 1, 2, 0, 1, 2}, got)
 	require.Equal(t, []int{2, 3}, dims)
 
 }